+++ workspaces/adoption-insights/packages/app-legacy/knip-report.md
@@ -4,13 +4,15 @@
 
 | Name         | Location          | Severity |
 | :----------- | :---------------- | :------- |
-| react-router | package.json:53:6 | error    |
+| react-router | package.json:55:6 | error    |
 
-## Unused devDependencies (4)
+## Unused devDependencies (6)
 
 | Name                        | Location          | Severity |
 | :-------------------------- | :---------------- | :------- |
-| @testing-library/user-event | package.json:62:6 | error    |
-| @backstage/test-utils       | package.json:57:6 | error    |
-| @testing-library/dom        | package.json:59:6 | error    |
-| cross-env                   | package.json:64:6 | error    |
+| @testing-library/user-event | package.json:65:6 | error    |
+| @backstage/test-utils       | package.json:60:6 | error    |
+| @axe-core/playwright        | package.json:59:6 | error    |
+| @testing-library/dom        | package.json:62:6 | error    |
+| @playwright/test            | package.json:61:6 | error    |
+| cross-env                   | package.json:67:6 | error    |
