+++ workspaces/adoption-insights/packages/app-legacy/src/App.tsx
@@ -50,7 +50,7 @@ import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
 import { RequirePermission } from '@backstage/plugin-permission-react';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { AdoptionInsightsPage } from '@red-hat-developer-hub/backstage-plugin-adoption-insights';
+import { AdoptionInsightsPage } from '@red-hat-developer-hub/backstage-plugin-adoption-insights/legacy';
 import { adoptionInsightsTranslations } from '@red-hat-developer-hub/backstage-plugin-adoption-insights/alpha';
 import { githubAuthApiRef } from '@backstage/core-plugin-api';
