+++ workspaces/adoption-insights/packages/app/src/App.tsx
@@ -18,8 +18,8 @@ import catalogPlugin from '@backstage/plugin-catalog/alpha';
 import userSettingsPlugin from '@backstage/plugin-user-settings/alpha';
 import adoptionInsightsPlugin, {
   adoptionInsightsTranslationsModule,
-} from '@red-hat-developer-hub/backstage-plugin-adoption-insights/alpha';
-import adoptionInsightsAnalyticsPlugin from '@red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights/alpha';
+} from '@red-hat-developer-hub/backstage-plugin-adoption-insights';
+import adoptionInsightsAnalyticsPlugin from '@red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights';
 import { navModule } from './modules/nav';
 import { signInModule } from './modules/signIn';
