+++ workspaces/adoption-insights/plugins/adoption-insights-common/knip-report.md
@@ -1,0 +1,8 @@
+# Knip report
+
+## Unlisted dependencies (2)
+
+| Name                                | Location       | Severity |
+| :---------------------------------- | :------------- | :------- |
+| @ianvs/prettier-plugin-sort-imports | .prettierrc.js | error    |
+| @spotify/prettier-config            | .prettierrc.js | error    |
