+++ workspaces/adoption-insights/plugins/adoption-insights/knip-report.md
@@ -1,14 +1,8 @@
 # Knip report
 
-## Unused dependencies (1)
-
-| Name             | Location          | Severity |
-| :--------------- | :---------------- | :------- |
-| @material-ui/lab | package.json:44:6 | error    |
-
 ## Unused devDependencies (2)
 
 | Name                    | Location          | Severity |
 | :---------------------- | :---------------- | :------- |
-| @backstage/core-app-api | package.json:59:6 | error    |
-| msw                     | package.json:66:6 | error    |
+| @backstage/core-app-api | package.json:82:6 | error    |
+| msw                     | package.json:90:6 | error    |
