+++ workspaces/adoption-insights/plugins/adoption-insights/package.json
@@ -2,18 +2,22 @@
   "name": "@red-hat-developer-hub/backstage-plugin-adoption-insights",
   "version": "0.9.1",
   "license": "Apache-2.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "src/index.tsx",
+  "types": "src/index.tsx",
   "exports": {
-    ".": "./src/index.ts",
-    "./alpha": "./src/alpha.tsx",
+    ".": "./src/index.tsx",
+    "./alpha": "./src/alpha.ts",
+    "./legacy": "./src/legacy.ts",
     "./adoption-insights-translations-module": "./src/adoptionInsightsTranslationsModuleExport.ts",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "alpha": [
-        "src/alpha.tsx"
+        "src/alpha.ts"
+      ],
+      "legacy": [
+        "src/legacy.ts"
       ],
       "adoption-insights-translations-module": [
         "src/adoptionInsightsTranslationsModuleExport.ts"
@@ -24,11 +28,7 @@
     }
   },
   "publishConfig": {
-    "access": "public",
-    "alpha": {
-      "types": "dist/alpha.d.ts",
-      "default": "dist/alpha.esm.js"
-    }
+    "access": "public"
   },
   "repository": {
     "type": "git",
@@ -47,6 +47,7 @@
   "sideEffects": false,
   "scripts": {
     "start": "backstage-cli package start",
+    "start:legacy": "backstage-cli package start --entrypoint dev/legacy.tsx",
     "build": "backstage-cli package build",
     "lint": "backstage-cli package lint",
     "test": "backstage-cli package test",
@@ -74,22 +75,26 @@
     "recharts": "^2.15.1"
   },
   "peerDependencies": {
-    "react": "^16.13.1 || ^17.0.0 || ^18.0.0"
+    "react": "^16.13.1 || ^17.0.0 || ^18.0.0",
+    "react-dom": "^16.13.1 || ^17.0.0 || ^18.0.0"
   },
   "devDependencies": {
     "@backstage/cli": "^0.36.3",
     "@backstage/config": "^1.3.8",
     "@backstage/core-app-api": "^1.20.2",
     "@backstage/dev-utils": "^1.1.24",
+    "@backstage/frontend-defaults": "^0.5.3",
     "@backstage/frontend-test-utils": "^0.6.1",
     "@backstage/plugin-catalog": "^2.0.6",
     "@backstage/test-utils": "^1.7.19",
+    "@backstage/ui": "^0.16.0",
     "@red-hat-developer-hub/backstage-plugin-theme": "^0.14.11",
     "@testing-library/jest-dom": "^6.0.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.0.0",
     "msw": "^1.0.0",
-    "react": "^16.13.1 || ^17.0.0 || ^18.0.0"
+    "react": "^16.13.1 || ^17.0.0 || ^18.0.0",
+    "react-dom": "^16.13.1 || ^17.0.0 || ^18.0.0"
   },
   "files": [
     "app-config.dynamic.yaml",
@@ -99,8 +104,9 @@
   "scalprum": {
     "name": "red-hat-developer-hub.backstage-plugin-adoption-insights",
     "exposedModules": {
-      "PluginRoot": "./src/index.ts",
-      "Alpha": "./src/alpha.tsx"
+      "PluginRoot": "./src/index.tsx",
+      "Legacy": "./src/legacy.ts",
+      "Alpha": "./src/alpha.ts"
     }
   }
 }
