+++ workspaces/adoption-insights/plugins/adoption-insights/report-legacy.api.md
@@ -1,0 +1,27 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-adoption-insights"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackstagePlugin } from '@backstage/core-plugin-api';
+import { IconComponent } from '@backstage/core-plugin-api';
+import { JSX as JSX_2 } from 'react/jsx-runtime';
+import { RouteRef } from '@backstage/core-plugin-api';
+
+// @public
+export const AdoptionInsightsIcon: IconComponent;
+
+// @public
+export const AdoptionInsightsPage: () => JSX_2.Element | null;
+
+// @public
+export const adoptionInsightsPlugin: BackstagePlugin<
+  {
+    root: RouteRef<undefined>;
+  },
+  {},
+  {}
+>;
+
+// (No @packageDocumentation comment for this package)
+```
