+++ workspaces/adoption-insights/plugins/adoption-insights/src/legacy.ts
@@ -13,4 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { adoptionInsightsTranslationsModule as default } from './index';
+
+/**
+ * Legacy frontend system API surface for the Adoption Insights plugin.
+ *
+ * @packageDocumentation
+ */
+
+import { unstable_ClassNameGenerator as ClassNameGenerator } from '@mui/material/className';
+
+ClassNameGenerator.configure(componentName => {
+  return componentName.startsWith('v5-')
+    ? componentName
+    : `v5-${componentName}`;
+});
+
+export * from './legacy/plugin';
