--- workspaces/adoption-insights/plugins/analytics-module-adoption-insights/report-alpha.api.md
@@ -1,37 +1,0 @@
-## API Report File for "@red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { AnalyticsImplementationFactory } from '@backstage/frontend-plugin-api';
-import { ExtensionBlueprintParams } from '@backstage/frontend-plugin-api';
-import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
-import { OverridableFrontendPlugin } from '@backstage/frontend-plugin-api';
-
-// @alpha (undocumented)
-const _default: OverridableFrontendPlugin<
-  {},
-  {},
-  {
-    'analytics:analytics-module-adoption-insights/analytics-module-adoption-insights': OverridableExtensionDefinition<{
-      kind: 'analytics';
-      name: 'analytics-module-adoption-insights';
-      config: {};
-      configInput: {};
-      output: ExtensionDataRef<
-        AnalyticsImplementationFactory<{}>,
-        'core.analytics.factory',
-        {}
-      >;
-      inputs: {};
-      params: <TDeps extends { [name in string]: unknown }>(
-        params: AnalyticsImplementationFactory<TDeps>,
-      ) => ExtensionBlueprintParams<AnalyticsImplementationFactory<{}>>;
-    }>;
-  }
->;
-export default _default;
-
-// (No @packageDocumentation comment for this package)
-```
