+++ workspaces/adoption-insights/plugins/analytics-module-adoption-insights/report-legacy.api.md
@@ -1,0 +1,38 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-analytics-module-adoption-insights"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { AnalyticsApi } from '@backstage/core-plugin-api';
+import { AnalyticsEvent } from '@backstage/frontend-plugin-api';
+import { AnalyticsEvent as AnalyticsEvent_2 } from '@backstage/core-plugin-api';
+import { AnalyticsImplementation } from '@backstage/frontend-plugin-api';
+import { ApiFactory } from '@backstage/frontend-plugin-api';
+import { Config } from '@backstage/config';
+import { ConfigApi } from '@backstage/core-plugin-api';
+import { IdentityApi } from '@backstage/core-plugin-api';
+import { IdentityApi as IdentityApi_2 } from '@backstage/frontend-plugin-api';
+
+// @public
+export class AdoptionInsightsAnalyticsApi
+  implements AnalyticsApi, AnalyticsImplementation
+{
+  captureEvent(event: AnalyticsEvent | AnalyticsEvent_2): Promise<void>;
+  static fromConfig(
+    config: ConfigApi,
+    options: {
+      identityApi: IdentityApi;
+    },
+  ): AdoptionInsightsAnalyticsApi;
+}
+
+// @public
+export const AdoptionInsightsAnalyticsApiFactory: ApiFactory<
+  AnalyticsApi,
+  AdoptionInsightsAnalyticsApi,
+  {
+    configApi: Config;
+    identityApi: IdentityApi_2;
+  }
+>;
+```
